seperate the main.cpp from proj and make the proj as lib

+++ ECSproject/ECS/SpriteComponent.h
@@ -121,7 +121,7 @@ public:
 	}
 
 	void OnInjuried() {
-		beginAttackTime = glfwGetTime();
+		beginAttackTime = (float) glfwGetTime();
 		_OnAttack = true;
 	}
